Refactor: pairing method is extracted from the tournament class, modules separated. TODO: define module responsibilities and refactor further
