+++ tournament/tournament.py
@@ -1,6 +1,5 @@
 import random
-from pathlib import Path
-from fighter import Fighter, fighter_from_str
+from tournament.fighter import Fighter, fighter_from_str
 from typing import Tuple, List
 
 def fight(f1: Fighter, f2: Fighter, result: Tuple[int, int]):
@@ -18,37 +17,9 @@ def hp(fighter: Fighter) -> int:
     # Function to sort fighters
     return fighter.hp
 
-
-def alreadyPlayed(player1: Fighter, player2: Fighter) -> bool:
-    """Have these two players already played in this tour?
-
-        A boolean indicating if this would be a rematch
-    """
-    return player1.played(player2) > 0 or player2.played(player1) > 0
-
-
-def decorate(filename):
-    """
-    We do not want to lose any of the data due to overwriting
-    :param filename:
-    :return:
-    """
-    if Path(filename).exists:
-        noext = filename[:filename.rfind('.')]
-        ext = filename[filename.rfind('.'):]
-        decorated = filename
-        i=1
-        while Path(decorated).exists():
-            decorated = noext + str(i) + ext
-            i += 1
-        return decorated
-    else:
-        return filename
-
-
 class Tournament:
 
-    def __init__(self, fighters: List[Fighter]=None, maxHP=12, fightCap=4):
+    def __init__(self, pairing_function, fighters: List[Fighter]=None, maxHP=12, fightCap=4):
 
         if fighters is not None:
             self.fighters = fighters
@@ -59,6 +30,10 @@ class Tournament:
         self.maxHP = maxHP
         self.fightCap = fightCap
         self.pairings = []
+        self.pairing_function = pairing_function
+
+    def make_pairs(self):
+        self.pairings = self.pairing_function(self.fighters)
 
     def update_fighters(self, name1: str, name2: str, score: Tuple[int, int]):
         """
@@ -141,36 +116,6 @@ class Tournament:
         results = [self.parse_result(res) for res in data]
         for res in results:
             self.update_fighters(*res)
-
-    def swissPairings(self):
-        """Returns a list of pairs of players for the next round of a match in this tour.
-
-        Assuming that there are an even number of players registered, each player
-        appears exactly once in the pairings.  Each player is paired with another
-        player with an equal or nearly-equal HP, that is, a player adjacent
-        to him or her in the standings.  Rematches are not allowed, so all pairings are new
-        (excluding situations when all pairs have matched, see Fighter.normalize_played )
-
-        Returns: a list of tuples of fighters
-        """
-
-        if len(self.fighters) % 2 != 0 or len(self.fighters) == 0:
-            raise ValueError("Number of fighters is {}, does not suit for pairing".format(len(self.fighters)))
-
-        standings = sorted(self.fighters, key=hp)
-
-        pairings = []
-
-        for i in range(0, len(standings), 2):
-            for j in range(i + 1, len(standings)):
-                if not alreadyPlayed(standings[i], standings[j]):
-                    # good pair found, swap into place then break the inner loop
-                    while j > i + 1:
-                        standings[j - 1], standings[j] = standings[j], standings[j - 1]
-                        j -= 1
-                    break
-            pairings.append((standings[i], standings[i + 1]))
-        self.pairings = pairings
 
     def remove(self):
         """
